requirements.txt
# Core
python-dotenv==1.0.0

# Storage
boto3==1.34.0
minio==7.2.3

# Databases
psycopg[binary]==3.2.3
pymongo==4.10.1
sqlalchemy==2.0.23

# Data processing
pandas==2.2.3
pyarrow==18.1.0

# Testing
pytest==7.4.3